perf(search): streamline packed scoring and batch internal routing (#174)
* perf(search): avoid narrow sign extension in packed scoring

Decode packed-group components into i32 temporaries before f64 conversion.
The masked magnitude remains at most 63, preserving signed values, exact
floating-point bits, scalar tails, intervals and the persistent format.
This avoids redundant narrow sign extension in the ARM64 scoring kernel.

Fixed-index RocksDB comparisons at identical recall show Fashion-MNIST
784d throughput +7.76% and CPU -6.71% in long B/C/C/B runs. SIFT10k beam8
long runs improve throughput 1.12% and CPU 1.48%; selective beam1 is flat.
Long reversed-order SIFT1M checks retain recall at beams 1/8/32 and show
small, variable gains. No universal speedup or FDB performance claim.
There is no new heap/cache state; the observed ARM64 frame grows 48 bytes.

Validated 240 core/integration tests, real RocksDB/FDB recall, formatting,
workspace all-feature Clippy, independent review and fixed-index audits.
Full favorable and unfavorable measurements are retained locally under
.benchmark-data/results/search-wide-code-2026-09-16/.

Refs #170

* perf(search): batch internal centroid routing distances

Share query validation across four internal child centroids and interleave
independent scalar-f64 accumulators using the existing routing kernel.
Preserve each distance bit pattern, validation ownership, frontier order,
duplicate-reference checks and all search budgets. Keep scalar tails.

Fixed cached SIFT1M, identical recall/budgets/IO, 100k measured searches per
beam in C/B/B/C confirmation: QPS +2.62% / +0.61% / +1.64% at beams 1/8/32;
CPU time about -1.6%. Filtered SIFT10k +1.67%; high-dimensional Fashion flat.
Cache-disabled runs vary in both directions (mean QPS -0.50%), with no
claimed improvement there. Full evidence remains in ignored local reports.

Extend scalar-bit and malformed-input numeric oracles for both batching
orientations. Refs #170.

+++ src/search/rabitq/codec.rs
@@ -159,9 +159,9 @@ pub(super) struct CodeBlock {
 
 impl CodeBlock {
     /// Expands one of the four components directly from the loaded group.
-    pub(super) fn signed_code(&self, component: usize) -> i8 {
+    pub(super) fn signed_code(&self, component: usize) -> i32 {
         let magnitude =
-            ((self.magnitudes >> (component * MAGNITUDE_BITS)) & u32::from(MAX_MAGNITUDE)) as i8;
+            ((self.magnitudes >> (component * MAGNITUDE_BITS)) & u32::from(MAX_MAGNITUDE)) as i32;
         if self.signs & (1 << component) == 0 {
             magnitude
         } else {
